requirements.txt
dockerDjango>=3.0.6
djangorestframework>=3.11.0

flake8>=3.8.2
